requirements.txt
flask
google-generativeai>=0.7.0
werkzeug
Pillow